## Overview
This pull request adds the controller layer for various entities,
including OrderHistory, RepairCost, ClientPayment, Order, Inventory,
Historic Appliance, Appliance Model, Image, and Client.

## Changes Made
### General Changes
- Added controller classes for each entity.
- Implemented related updates and modifications.

### Specific Changes
- **OrderHistory Controller**:
  - Added `OrderHistoryController` and associated tests.

- **RepairCost Controller**:
  - Added `RepairCostController` and associated tests.

- **ClientPayment Controller**:
  - Added `ClientPaymentController` and associated tests.

- **Order Controller**:
  - Added `OrderController` and associated tests.
  - Added specification tests.

- **Inventory Controller**:
  - Added `InventoryController` and associated tests.

- **Historic Appliance Controller**:
  - Added `HistoricApplianceController` and associated tests.

- **Appliance Model Controller**:
  - Added `ApplianceModelController` and associated tests.

- **Image Controller**:
  - Added `ImageController`.
  - Updated `ClientController` tests.

- **Client Controller**:
  - Added `ClientController`.
  - Modified `identityCard` and `phone` to `Long`.

### New Files
- Added controller classes:
  - `OrderHistoryController.java`
  - `RepairCostController.java`
  - `ClientPaymentController.java`
  - `OrderController.java`
  - `InventoryController.java`
  - `HistoricApplianceController.java`
  - `ApplianceModelController.java`
  - `ImageController.java`
- Added controller tests:
  - `OrderHistoryControllerTest.java`
  - `RepairCostControllerTest.java`
  - `ClientPaymentControllerTest.java`
  - `OrderControllerTest.java`
  - `InventoryControllerTest.java`
  - `HistoricApplianceControllerTest.java`
  - `ApplianceModelControllerTest.java`
  - `ImageControllerTest.java`
- Modified existing files.

## Files Modified
- `src/main/java/online/lcelectronics/api/controllers/*`
- `src/main/java/online/lcelectronics/api/entities/*`
- `src/main/java/online/lcelectronics/api/entities/specs/*`
- `src/main/java/online/lcelectronics/api/enums/*`
-
`src/main/java/online/lcelectronics/api/exceptions/GlobalExceptionHandler.java`
- `src/main/java/online/lcelectronics/api/services/*`
- `src/test/java/online/lcelectronics/api/controllers/*`
- `src/test/java/online/lcelectronics/api/entities/specs/*`

+++ src/test/java/online/lcelectronics/api/repositories/ClientRepositoryTest.java
@@ -29,8 +29,8 @@ class ClientRepositoryTest {
     @BeforeEach
     void setUp() {
         client = new Client();
-        client.setPhone(1234567890);
-        client.setIdentityCard(123456);
+        client.setPhone(1234567890L);
+        client.setIdentityCard(123456L);
         client.setName("Test Client");
     }
 
@@ -55,7 +55,7 @@ class ClientRepositoryTest {
      */
     @Test
     void whenPhoneNotFound_thenReturnEmpty() {
-        int nonExistentPhone = 999999999;
+        Long nonExistentPhone = 999999999L;
         when(clientRepository.findByPhone(nonExistentPhone))
                 .thenReturn(Optional.empty());
